% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_background_data.R
\name{extract_kegg_ids}
\alias{extract_kegg_ids}
\title{Extract KEGG compound IDs from pathway data}
\usage{
extract_kegg_ids(PathwayVsMetabolites)
}
\arguments{
\item{PathwayVsMetabolites}{Data frame containing a
\code{Metabolites} column with comma-separated KEGG compound IDs.}
}
\value{
A character vector containing unique KEGG compound IDs.
}
\description{
Extracts unique KEGG compound IDs from the \code{Metabolites}
column of a KEGG pathway-metabolite table.
}
\examples{

PathwayVsMetabolites <- fetch_kegg_pathway_metabolites(organism = "hsa")
kegg_ids <- extract_kegg_ids(PathwayVsMetabolites)
head(kegg_ids)


}
