% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_background_data.R
\name{fetch_kegg_chebi}
\alias{fetch_kegg_chebi}
\title{Fetch KEGG-ChEBI mappings}
\usage{
fetch_kegg_chebi(kegg_ids)
}
\arguments{
\item{kegg_ids}{Character vector of KEGG compound IDs.}
}
\value{
A data frame with columns \code{KEGG} and \code{ChEBI}.
}
\description{
Downloads the KEGG-to-ChEBI conversion table and filters
it to the supplied KEGG compound IDs.
}
\examples{

kegg_chebi <- fetch_kegg_chebi(c("C00022", "C00031"))
head(kegg_chebi)


}
