% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_plot.R
\name{create_heatmap_plot}
\alias{create_heatmap_plot}
\title{Create Enrichment Heatmap}
\usage{
create_heatmap_plot(
  enrichment_results,
  PathwayVsMetabolites,
  inputMetabolites,
  kegg_lookup = NULL,
  min_pathways = 1,
  min_metabolites = 1,
  base_font_size = 14
)
}
\arguments{
\item{enrichment_results}{Data frame from perform_enrichment_analysis().}

\item{PathwayVsMetabolites}{A data frame with pathway-metabolite associations.}

\item{inputMetabolites}{A character vector of metabolite IDs.}

\item{kegg_lookup}{Optional data frame for KEGG ID to name mapping.}
}
\value{
A ComplexHeatmap object showing enrichment significance.
}
\description{
Generates a heatmap showing metabolite-pathway enrichment significance.
}
\examples{
# Create comprehensive example data
set.seed(123)
enrichment_results <- data.frame(
  Pathway = c("Glycolysis / Gluconeogenesis", "Citrate cycle (TCA cycle)",
              "Pentose phosphate pathway", "Pyruvate metabolism",
              "Amino sugar and nucleotide sugar metabolism",
              "Glycine, serine and threonine metabolism",
              "Cysteine and methionine metabolism"),
  Log_P_value = c(7.2, 5.8, 4.9, 4.3, 3.8, 3.5, 3.2),
  P_value = 10^(-c(7.2, 5.8, 4.9, 4.3, 3.8, 3.5, 3.2)),
  Adjusted_P_value = 10^(-c(6.5, 5.2, 4.4, 3.9, 3.5, 3.2, 2.9))
)

PathwayVsMetabolites <- data.frame(
  Pathway = c("Glycolysis / Gluconeogenesis", "Citrate cycle (TCA cycle)",
              "Pentose phosphate pathway", "Pyruvate metabolism",
              "Amino sugar and nucleotide sugar metabolism",
              "Glycine, serine and threonine metabolism",
              "Cysteine and methionine metabolism"),
  Metabolites = c("C00031,C00022,C00074,C00036,C00103,C00197,C00186,C00221,C01172",
                  "C00024,C00036,C00042,C00122,C00149,C00158,C00417,C05379",
                  "C00117,C00231,C00279,C00345,C01172,C01236,C01218",
                  "C00022,C00024,C00036,C00074,C00122,C00149,C00232",
                  "C00031,C00095,C00103,C00121,C00140,C00208,C00267",
                  "C00037,C00048,C00065,C00143,C00258,C00300,C00497",
                  "C00021,C00073,C00094,C00155,C00283,C00542,C00957")
)

inputMetabolites <- c("C00031", "C00022", "C00074", "C00036", "C00103", "C00197", 
                     "C00186", "C00221", "C00024", "C00042", "C00122", "C00149", 
                     "C00158", "C00117", "C00231", "C00279", "C00037", "C00048")

kegg_lookup <- data.frame(
  kegg_id = c("C00031", "C00022", "C00074", "C00036", "C00103", "C00197", 
              "C00186", "C00221", "C00024", "C00042", "C00122", "C00149", 
              "C00158", "C00117", "C00231", "C00279", "C00037", "C00048",
              "C00417", "C05379", "C00345", "C01172", "C01236", "C00065",
              "C00143", "C00258", "C00021", "C00073", "C00094"),
  name = c("D-Glucose", "L-Lactate", "Oxaloacetate", "Citrate", "D-Fructose 6-phosphate",
           "3-Phospho-D-glyceroyl phosphate", "L-Aspartate", "Oxoglutaric acid",
           "Acetyl-CoA", "Oxaloacetic acid", "L-Glutamate", "Succinyl-CoA",
           "L-Malate", "D-Ribose 5-phosphate", "D-Xylulose 5-phosphate",
           "Sedoheptulose 7-phosphate", "Glycine", "L-Serine",
           "Succinate", "Isocitrate", "Glyceraldehyde 3-phosphate",
           "D-Fructose 1,6-bisphosphate", "D-Erythrose 4-phosphate",
           "L-Threonine", "L-Cysteine", "O-Acetyl-L-serine",
           "L-Cystathionine", "L-Homocysteine", "S-Adenosyl-L-methionine")
)

# Create enrichment heatmap
heatmap_plot <- create_heatmap_plot(enrichment_results, PathwayVsMetabolites, 
                                   inputMetabolites, kegg_lookup)
# Note: ComplexHeatmap objects need to be drawn explicitly
if (!is.null(heatmap_plot)) {
  ComplexHeatmap::draw(heatmap_plot)
}

}
