% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_membership_plot.R
\name{create_membership_plot}
\alias{create_membership_plot}
\title{Create Pathway Membership Plot}
\usage{
create_membership_plot(
  PathwayVsMetabolites,
  inputMetabolites,
  kegg_lookup = NULL
)
}
\arguments{
\item{PathwayVsMetabolites}{A data frame with pathway-metabolite associations.}

\item{inputMetabolites}{A character vector of metabolite IDs.}

\item{kegg_lookup}{Optional data frame for KEGG ID to name mapping.}
}
\value{
A ComplexHeatmap object showing pathway membership.
}
\description{
Generates a heatmap showing pathway membership for metabolites.
}
\examples{
# Create comprehensive example data with matching metabolite IDs
inputMetabolites <- c("C00031", "C00022", "C00074", "C00036", "C00103", 
                     "C00197", "C00186", "C00221", "C00024", "C00042",
                     "C00122", "C00149", "C00158", "C00117", "C00231",
                     "C00279", "C00037", "C00048", "C00065", "C00143")

PathwayVsMetabolites <- data.frame(
  Pathway = c("Glycolysis/Gluconeogenesis", "Citrate cycle (TCA cycle)", 
              "Pentose phosphate pathway", "Pyruvate metabolism",
              "Amino sugar and nucleotide sugar metabolism",
              "Glycine, serine and threonine metabolism",
              "Cysteine and methionine metabolism",
              "Valine, leucine and isoleucine degradation"),
  Metabolites = c("C00031,C00022,C00074,C00036,C00103,C00197,C00186,C00221",
                  "C00024,C00036,C00042,C00122,C00149,C00158,C00417",
                  "C00117,C00231,C00279,C00345,C01172",
                  "C00022,C00024,C00036,C00074,C00122",
                  "C00031,C00095,C00103,C00121,C00140",
                  "C00037,C00048,C00065,C00143,C00258",
                  "C00021,C00073,C00094,C00155,C00283",
                  "C00123,C00183,C00233,C00407,C00979")
)

# Create KEGG lookup with matching IDs
kegg_lookup <- data.frame(
  kegg_id = c("C00031", "C00022", "C00074", "C00036", "C00103", 
              "C00197", "C00186", "C00221", "C00024", "C00042",
              "C00122", "C00149", "C00158", "C00117", "C00231",
              "C00279", "C00037", "C00048", "C00065", "C00143",
              "C00258", "C00021", "C00073", "C00094", "C00155"),
  name = c("D-Glucose", "L-Lactate", "Oxaloacetate", "Citrate", 
           "D-Fructose 6-phosphate", "3-Phospho-D-glyceroyl phosphate",
           "L-Aspartate", "Oxoglutaric acid", "Acetyl-CoA", 
           "Oxaloacetic acid", "L-Glutamate", "Succinyl-CoA",
           "L-Malate", "D-Ribose 5-phosphate", "D-Xylulose 5-phosphate",
           "Sedoheptulose 7-phosphate", "Glycine", "L-Serine",
           "L-Threonine", "L-Cysteine", "O-Acetyl-L-serine",
           "L-Cystathionine", "L-Homocysteine", "S-Adenosyl-L-methionine", "Coenzyme A")
)

# Create membership plot
plot <- create_membership_plot(PathwayVsMetabolites, inputMetabolites, kegg_lookup)
# Note: ComplexHeatmap objects need to be drawn explicitly
if (!is.null(plot)) {
  ComplexHeatmap::draw(plot)
}
}
