% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_plot.R
\name{create_network_plot}
\alias{create_network_plot}
\title{Create Metabolite-Pathway Network}
\usage{
create_network_plot(inputMetabolites, PathwayVsMetabolites, kegg_lookup = NULL)
}
\arguments{
\item{inputMetabolites}{A character vector of metabolite IDs.}

\item{PathwayVsMetabolites}{A data frame with pathway-metabolite associations.}

\item{kegg_lookup}{Optional data frame for KEGG ID to name mapping.}
}
\value{
A ggraph object showing the metabolite-pathway network.
}
\description{
Generates a metabolite-pathway interaction network visualization.
}
\examples{
# Create comprehensive example data
inputMetabolites <- paste0("C", sprintf("\%05d", 1:25))

PathwayVsMetabolites <- data.frame(
  Pathway = c("Glycolysis", "TCA cycle", "Pentose phosphate", 
              "Amino acid metabolism", "Lipid metabolism",
              "Nucleotide metabolism", "Oxidative phosphorylation"),
  Metabolites = c("C00031,C00022,C00118,C00197,C00074,C00036,C00186,C00221,C00089,C00103",
                  "C00024,C00036,C00042,C00122,C00149,C00158,C00417,C05379",
                  "C00117,C00231,C00279,C00345,C01172,C01236",
                  "C00025,C00026,C00041,C00049,C00064,C00135",
                  "C00083,C00162,C00249,C00422,C01205",
                  "C00106,C00112,C00144,C00212,C00262",
                  "C00009,C00013,C00080,C01345,C05345")
)

# Create KEGG lookup
kegg_lookup <- data.frame(
  kegg_id = paste0("C", sprintf("\%05d", 1:30)),
  name = c("Glucose", "Lactate", "Pyruvate", "Alanine", "Glutamate",
           "Citrate", "Succinate", "Malate", "Aspartate", "Glutamine",
           "Acetyl-CoA", "Oxaloacetate", "Alpha-ketoglutarate", "Fumarate",
           "Isocitrate", "Ribose-5P", "Glycerol-3P", "Dihydroxyacetone-P",
           "Fructose-6P", "Glucose-6P", "Phosphoenolpyruvate", "ATP", "ADP",
           "NAD+", "NADH", "Coenzyme A", "Acetate", "Butyrate", "Propionate", "Glycine")
)

# Create network plot
plot <- create_network_plot(inputMetabolites, PathwayVsMetabolites, kegg_lookup)
plot

}
