% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_background_data.R
\name{fetch_kegg_reactome}
\alias{fetch_kegg_reactome}
\title{Build KEGG-Reactome reaction mappings}
\usage{
fetch_kegg_reactome(kegg_ids, species = "Homo sapiens")
}
\arguments{
\item{kegg_ids}{Character vector of KEGG compound IDs.}

\item{species}{Character. Reactome species name.}
}
\value{
A data frame containing KEGG-to-ChEBI-to-Reactome mappings.
}
\description{
Links KEGG compounds to Reactome reactions through ChEBI identifiers.
}
\examples{
\donttest{
reactome_df <- fetch_kegg_reactome(
    kegg_ids = c("C00022", "C00031"),
    species = "Homo sapiens"
)
head(reactome_df)
}

}
