% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_background_data.R
\name{get_cached_file}
\alias{get_cached_file}
\title{Download and cache a remote file}
\usage{
get_cached_file(url)
}
\arguments{
\item{url}{Character. URL of the remote file.}
}
\value{
Character path to the cached file.
}
\description{
Downloads a remote resource and stores it in the EnrichMet cache.
Existing cached resources are reused automatically.
}
\examples{
\donttest{
path <- get_cached_file(
    "https://zenodo.org/api/records/17819145/files/summary_stat.csv/content"
)
file.exists(path)
}

}
